\name{SeqCNVCBS-package}
\alias{SeqCNVCBS-package}
\alias{SeqCNVCBS}
\docType{package}
\title{
Scan Statistics CNV detection using sequencing data
}
\description{
CNV detection using matched case-control sequencing read data. It gives a number of scan statistics for the detection of rate differences between two non-homogeneous Poisson Processes, and modified BIC model selection.
}
\details{
\tabular{ll}{
Package: \tab SeqCNVCBS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-04-27\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Jeremy J. Shen\cr
Nancy R. Zhang

Maintainer: Jeremy J. Shen <jqshen@stanford.edu>
~~ The author and/or maintainer of the package ~~
}

\keyword{ package }
